fix: lint with pre-commit
