+++ github_stars_by_topic/__init__.py
@@ -1,1 +1,3 @@
 __version__ = "0.0.0"
+
+CACHE_PATH_READMES = "cache"
