+++ setup.cfg
@@ -7,8 +7,6 @@ long_description_content_type = text/markdown
 url = https://github.com/lorey/github-stars-by-topic
 author = Karl Lorey
 author_email = git@karllorey.com
-# license =
-# license_file = LICENSE
 classifiers =
     Development Status :: 3 - Alpha
     Programming Language :: Python :: 3
@@ -20,39 +18,26 @@ classifiers =
 
 [options]
 packages = find:
-python_requires = >=3.7
 install_requires =
-    beautifulsoup4==4.11.1
     Markdown==3.3.6
-    numpy==1.21.6
     PyGithub==1.55
-    scikit-learn==1.10.2
-
+    beautifulsoup4==4.11.1
+    numpy==1.21.6
+    scikit-learn==1.0.2
+python_requires = >=3.7
 packages_data =
     github_stars_by_topics = py.typed
 
 [options.entry_points]
 console_scripts =
-    ghst = .main:main
+    ghst = github_stars_by_topics.main:main
 
 [options.extras_require]
 all =
-    %(example)s
     %(dev)s
-example =
-    scipy==1.5.1
-    soupsieve==2.0.1
-    threadpoolctl==2.1.0
-    urllib3==1.25.9
-    wrapt==1.12.1
-    certifi==2020.6.20
-    chardet==3.0.4
-    Deprecated==1.2.10
-    idna==2.10
-    joblib==0.16.0
-    PyJWT==1.7.1
-    requests==2.24.0
+    %(example)s
 dev =
+    %(types)s
     black
     flake8
     isort
@@ -60,11 +45,23 @@ dev =
     pre-commit
     setuptools>=46.4.0
     types-setuptools
-    %(types)s
+example =
+    Deprecated==1.2.10
+    PyJWT==1.7.1
+    certifi==2020.6.20
+    chardet==3.0.4
+    idna==2.10
+    joblib==0.16.0
+    requests==2.24.0
+    scipy==1.5.1
+    soupsieve==2.0.1
+    threadpoolctl==2.1.0
+    urllib3==1.25.9
+    wrapt==1.12.1
 types =
-    types-setuptools
-    types-beautifulsoup4
     types-Markdown
+    types-beautifulsoup4
+    types-setuptools
 
 [isort]
 profile = black
